Alterar estacioamentos

+++ app/Http/Controllers/Parking/ParkingController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Parking;
 
+use App\Dtos\Error\ErrorDTO;
 use App\Dtos\Parking\OutputParkingDTO;
 use App\Dtos\Parking\ParkingDTO;
 use App\Http\Controllers\Controller;
@@ -41,13 +42,23 @@ class ParkingController extends Controller
 
     public function show(string $id)
     {
-        $parking = $this->parking::with('cars')->findOrFail($id);
+        $parking = $this->parking::with('cars')->find($id);
+
+        if(empty($parking)){
+            return $this->outputErrorResponse($id);
+        }
 
         return $this->outputResponse($parking);
     }
 
     public function update(Request $request, string $id)
     {
+        $parking = $this->parking::find($id);
+
+        if(empty($parking)){
+            return $this->outputErrorResponse($id);
+        }
+
         $dto = new ParkingDTO(
             ...$request->only([
                 'name',
@@ -56,8 +67,6 @@ class ParkingController extends Controller
             ])
         );
 
-        $parking = $this->parking::findOrFail($id);
-
         $parking->update($dto->toArray());
 
         return $this->outputResponse($parking);
@@ -65,7 +74,11 @@ class ParkingController extends Controller
 
     public function destroy(string $id)
     {
-        $parking = $this->parking::findOrFail($id);
+        $parking = $this->parking::find($id);
+
+        if(empty($parking)){
+            return $this->outputErrorResponse($id);
+        }
 
         $parking->delete();
 
@@ -73,16 +86,22 @@ class ParkingController extends Controller
     }
 
     private function outputResponse(Parking $parking){
+
         $output =  new OutputParkingDTO(
             $parking['id'],
             $parking['name'],
             $parking['numberOfVacancies'],
             $parking['active'],
             $parking['created_at'],
-            // $parking['cars'],
-            // $parking['employees'],
+            $parking['cars'],
+            $parking['employees'],
         );
 
         return response()->json($output->toArray());
+    }
+
+    private function outputErrorResponse(int $id){
+        $error = new ErrorDTO("Registro {$id} não encontrado", Response::HTTP_NOT_FOUND);
+        return response()->json($error->toArray(), Response::HTTP_NOT_FOUND);
     }
 }
